# API/swagger.yaml
...
swagger: "2.0"
info:
  description: "Student Service API"
  version: "1.0.0"
  title: "Student"
  termsOfService: ""
  contact:
    name: "someName"
    url: "http://www.staffshift.com"
    email: "info@staffshift.com"
host: "localhost:3000"
basePath: "/"
...
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /admin/login:
    post:
      tags:
        - Admin
      parameters:
        - name: loginPayload
          in: body
          required: true
          schema:
            $ref: "#/definitions/admin_login_payload"
      summary: "admin login"
      x-swagger-router-controller: admin-swagger
      operationId: adminLogin
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: "An admin is Logged in"
          schema:
            $ref: "#/definitions/admin"
        404:
          $ref: "#/responses/admin_credentials_error"
        default:
          $ref: "#/responses/default_server_error"
  /admin/technology/add-technology:
    post:
      tags:
        - Admin
      parameters:
        - name: technologyPayload
          in: body
          required: true
          schema:
            $ref: "#/definitions/technology_payload"
      summary: "add Technology"
      x-swagger-router-controller: admin-swagger
      operationId: addTechnology
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        201:
          description: "Technology is added"
          schema:
            $ref: "#/definitions/technology_response_payload"
        400:
          $ref: "#/responses/technology_error"
        default:
          $ref: "#/responses/default_server_error"
  /admin/technology/get-technology:
    get:
      tags:
        - Admin
      summary: "get technology"
      x-swagger-router-controller: admin-swagger
      operationId: getAllTechnology
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: "return all technology"
          schema:
            $ref: "#/definitions/allTechnology_payload"
        400:
          $ref: "#/responses/admin_credentials_error"
        default:
          $ref: "#/responses/default_server_error"
  /admin/technology/delete-technology/{technology_id}:
    delete:
      parameters:
        - name: technology_id
          in: path
          required: true
          type: string
      tags:
        - Admin
      summary: "delete technology"
      x-swagger-router-controller: admin-swagger
      operationId: deleteTechnology
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: "technology is deleted"
          schema:
            $ref: "#/definitions/deleteTechnology_response"
        404:
          $ref: "#/responses/admin_credentials_error"
        default:
          $ref: "#/responses/default_server_error"
  /admin/topic/add-topic:
    post:
      tags:
        - Admin
      parameters:
        - name: topicPayload
          in: body
          required: true
          schema:
            $ref: "#/definitions/topic_payload"
      summary: "add Topic"
      x-swagger-router-controller: admin-swagger
      operationId: addTopic
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        201:
          description: "Topic is added"
          schema:
            $ref: "#/definitions/topic_response_payload"
        400:
          $ref: "#/responses/topic_error"
        default:
          $ref: "#/responses/default_server_error"
  /admin/topic/get-topic:
    get:
      tags:
        - Admin
      summary: "get All Topic"
      x-swagger-router-controller: admin-swagger
      operationId: getAllTopic
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: "return all topic"
          schema:
            $ref: "#/definitions/allTopic_payload"
        400:
          $ref: "#/responses/topic_error"
        default:
          $ref: "#/responses/default_server_error"
  /admin/document/add-document:
    post:
      tags:
        - Admin
      x-swagger-router-controller: admin-swagger
      operationId: addDocument
      summary: "add documentation"
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: formData
          name: technology
          required: true
          type: string
        - in: formData
          name: topic
          required: true
          type: string
        - in: formData
          name: selectedDocument
          required: false
          type: file
        - in: formData
          name: selectedLink
          required: false
          type: string
      responses:
        201:
          description: "document is added"
          schema:
            $ref: "#/definitions/document_response_payload"
...
definitions:
...
  server_error_model:
    type: object
    required:
      - message
    properties:
      message:
        type: string
      exception:
        $ref: "#/definitions/exception_model"
  exception_model:
    type: object
...
  common_validation_error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: string
        enum:
          - PATTERN
          - REQUIRED
          - MIN_LENGTH
          - DATA_EXIST
          - DATA_NOT_EXIST
      message:
        type: string
  admin_payload:
    type: object
    required:
      - name
      - emailId
      - password
    properties:
      name:
        type: string
        minLength: 3
      emailId:
        type: string
        default: abc@gmail.com
        pattern: ^([\w.%+-]+)@([\w-]+\.)+([\w]{2,})$
      password:
        type: string
        minLength: 3
  admin_login_payload:
    type: object
    required:
      - emailId
      - password
    properties:
      emailId:
        type: string
        default: abc@gmail.com
        pattern: ^([\w.%+-]+)@([\w-]+\.)+([\w]{2,})$
      password:
        type: string
        minLength: 8
  mongoose_default:
    type: object
    required:
      - _id
      - __v
    properties:
      _id:
        type: object
        pattern: ^[0-9a-fA-F]{24}$
      __v:
        type: number
  token_response:
    type: object
    required:
      - token
    properties:
      token:
        type: string
  admin:
    allOf:
      - $ref: "#/definitions/admin_payload"
      - $ref: "#/definitions/mongoose_default"
      - $ref: "#/definitions/token_response"
  technology_payload:
    type: object
    required:
      - technology
      - hours
    properties:
      technology:
        type: string
        default: node js
        minLength: 5
      hours:
        type: number
        minLength: 2
  technology_response_payload:
    allOf:
      - $ref: "#/definitions/technology_payload"
      - $ref: "#/definitions/mongoose_default"
  allTechnology_payload:
    type: object
    properties:
      data:
        type: array
        items:
          - $ref: "#/definitions/technology_response_payload"
  deleteTechnology_response:
    type: object
    required:
      - message
    properties:
      message:
        type: string
  topic_payload:
    type: object
    required:
      - technology
      - topic
    properties:
      technology:
        type: string
        minLength: 3
      topic:
        type: string
        default: node js syntax
        minLength: 5
  topic_response_payload:
    allOf:
      - $ref: "#/definitions/topic_payload"
      - $ref: "#/definitions/mongoose_default"
  allTopic_payload:
    type: object
    properties:
      data:
        type: array
        items:
          - $ref: "#/definitions/topic_response_payload"
  document_payload:
    type: object
    required:
      - technology
      - topic
      - selectedDocument
    properties:
      technology:
        type: string
      topic:
        type: string
      selectedDocument:
        type: object
  document_response_payload:
    allOf:
      - $ref: "#/definitions/document_payload"
      - $ref: "#/definitions/mongoose_default"
responses:
...
  default_server_error:
    description: "unexpected error"
    schema:
      $ref: "#/definitions/server_error_model"
...
  admin_credentials_error:
    description: "error on create student"
    schema:
      $ref: "#/definitions/common_validation_error"
...
  technology_error:
    description: "error on create technology"
    schema:
      $ref: "#/definitions/common_validation_error"
  topic_error:
    description: "error on create topic"
    schema:
      $ref: "#/definitions/common_validation_error"
...